+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# 0.0.2
+# 0.0.2 (13-07-2014)
 
-* More options to build a stateful object
+* Improved documentation (see README)
+* Add support for same transition applied from different states
+* Add support for ignoring invalid transitions
+* Throw error if unknown initial state
 
 # 0.0.1 (01-05-2014)